full application refactored
