+++ package.json
@@ -15,6 +15,7 @@
     "request": "^2.88.2"
   },
   "devDependencies": {
+    "eslint": "^8.52.0",
     "nodemon": "^3.0.1"
   }
 }
